fixing unit tests, still 17 failing
