+++ tests/icone_translator_test.py
@@ -186,102 +186,6 @@ def test_parse_incident_no_direction():
     assert test_feature == None
 
 
-# --------------------------------------------------------------------------------Unit test for parse_polyline function--------------------------------------------------------------------------------
-def test_parse_icone_polyline_valid_data():
-    test_polyline = "34.8380671,-114.1450650,34.8380671,-114.1450650"
-    test_coordinates = icone_translator.parse_icone_polyline(
-        test_polyline)
-    valid_coordinates = [
-        [
-            -114.145065,
-            34.8380671
-        ],
-        [
-            -114.145065,
-            34.8380671
-        ]
-    ]
-    assert test_coordinates == valid_coordinates
-
-
-def test_parse_polyline_null_parameter():
-    test_polyline = None
-    test_coordinates = icone_translator.parse_icone_polyline(
-        test_polyline)
-    expected_coordinates = None
-    assert test_coordinates == expected_coordinates
-
-
-def test_parse_polyline_invalid_data():
-    test_polyline = 'invalid'
-    test_coordinates = icone_translator.parse_icone_polyline(
-        test_polyline)
-    expected_coordinates = []
-    assert test_coordinates == expected_coordinates
-
-
-def test_parse_polyline_invalid_coordinates():
-    test_polyline = 'a,b,c,d'
-    test_coordinates = icone_translator.wzdx_translator.parse_polyline(
-        test_polyline)
-    expected_coordinates = []
-    assert test_coordinates == expected_coordinates
-
-
-# --------------------------------------------------------------------------------Unit test for validate_incident function--------------------------------------------------------------------------------
-def test_validate_incident_valid_data():
-    test_valid_output = {
-        '@id': 'U13631595_202012160845',
-        'updatetime': '2020-12-16T17:18:00Z',
-        'starttime': '2020-12-07T14:18:00Z',
-        'description': 'Road constructions are going on',
-        'creationtime': '2020-12-13T14:18:00Z',
-        'location': {
-            'polyline': '34.8380671,-114.1450650,34.8380671,-114.1450650',
-            'street': 'I-70 N'
-        }
-    }
-    assert icone_translator.validate_incident(test_valid_output) == True
-
-
-def test_validate_incident_missing_required_field_description():
-    test_valid_output = {
-        '@id': 'U13631595_202012160845',
-        'updatetime': '2020-12-16T17:18:00Z',
-        'starttime': '2020-12-07T14:18:00Z',
-        'creationtime': '2020-12-13T14:18:00Z',
-        'location': {
-            'polyline': '34.8380671,-114.1450650,34.8380671,-114.1450650'
-        }
-    }
-    assert icone_translator.validate_incident(test_valid_output) == False
-
-
-def test_validate_incident_invalid_start_time():
-    test_valid_output = {
-        '@id': 'U13631595_202012160845',
-        'updatetime': '2020-12-16T17:18:00Z',
-        'starttime': 'dsafsaf',
-        'description': 'Road constructions are going on',
-        'creationtime': '2020-12-13T14:18:00Z',
-        'location': {
-            'polyline': '34.8380671,-114.1450650,34.8380671,-114.1450650',
-            'street': 'I-70 N'
-        }
-    }
-    assert icone_translator.validate_incident(test_valid_output) == False
-
-
-def test_validate_incident_invalid():
-    test_valid_output = 'invalid output'
-    assert icone_translator.validate_incident(test_valid_output) == False
-
-
-def test_validate_incident_no_data():
-    test_valid_output = None
-    assert icone_translator.validate_incident(test_valid_output) == False
-
-
 # --------------------------------------------------------------------------------Unit test for get_vehicle_impact function--------------------------------------------------------------------------------
 def test_get_vehicle_impact_some_lanes_closed():
     test_description = "Roadwork - Lane Closed, MERGE LEFT [Trafficade, iCone]"
@@ -597,39 +501,9 @@ def test_wzdx_creator_valid_and_invalid(mockuuid):
     with time_machine.travel(datetime(2021, 4, 13, 0, 0, 0)):
         test_wzdx = icone_translator.wzdx_creator(icone_obj)
     assert expected_wzdx == test_wzdx
-# --------------------------------------------------------------------------------unit test for parse_direction_from_street_name function--------------------------------------------------------------------------------
 
-
-def test_parse_direction_from_street_name_southbound():
-    test_road_name = 'I-75 S'
-    output_direction = icone_translator.parse_direction_from_street_name(
-        test_road_name)
-    assert output_direction == 'southbound'
-
-
-def test_parse_direction_from_street_name_northbound():
-    test_road_name = 'I-75 NB'
-    output_direction = icone_translator.parse_direction_from_street_name(
-        test_road_name)
-    assert output_direction == 'northbound'
-
-
-def test_parse_direction_from_street_name_eastbound():
-    test_road_name = 'I-75 EB'
-    output_direction = icone_translator.parse_direction_from_street_name(
-        test_road_name)
-    assert output_direction == 'eastbound'
-
-
-def test_parse_direction_from_street_name_westbound():
-    test_road_name = 'I-75 W'
-    output_direction = icone_translator.parse_direction_from_street_name(
-        test_road_name)
-    assert output_direction == 'westbound'
 
 # --------------------------------------------------------------------------------unit test for parse_icone_sensor function--------------------------------------------------------------------------------
-
-
 def test_parse_icone_sensor():
     valid_description = {'type': 'iCone', 'id': '#4', 'location': [41.3883260, -81.9707500], 'radar': {
         'average_speed': 63.52, 'std_dev_speed': 7.32, 'timestamp': '2020-08-21T15:55:00Z'}}
