+++ tests/raw_to_standard/navjoy_568_test.py
@@ -1,0 +1,396 @@
+from wzdx.raw_to_standard import navjoy_568
+from tests.raw_to_standard import navjoy_translator_test_expected_results as expected_results
+
+
+# --------------------------------------------------------------------------------Unit test for validate_closure function--------------------------------------------------------------------------------
+def test_validate_closure_valid_data():
+    event = {
+        "sys_gUid": "Form568-cb0fdaf0-c27a-4bef-aabd-442615dfb2d6",
+        "data": {
+            "descriptionForProject": "Bridge Repairs",
+            "srzmap": [
+                {
+                    "type": "LineString",
+                    "coordinates": [
+                        [
+                            [
+                                -103.17130040113868,
+                                40.625392709715676
+                            ],
+                            [
+                                -103.17889641706886,
+                                40.61979008921054
+                            ]
+                        ]
+                    ],
+                },
+            ],
+            "streetNameFrom": "US-34",
+            "direction": "eastbound",
+            "requestedTemporarySpeed": "45",
+            "workStartDate": "2021-08-09T13:00:00.000Z",
+            "workEndDate": "2021-09-24T23:00:00.000Z",
+            "reductionJustification": "Crews roadside.",
+        }
+    }
+    assert navjoy_568.validate_closure(event) == True
+
+
+def test_validate_closure_missing_required_field_description():
+    event = {
+        "sys_gUid": "Form568-cb0fdaf0-c27a-4bef-aabd-442615dfb2d6",
+        "data": {
+            "srzmap": [
+                {
+                    "type": "LineString",
+                    "coordinates": [
+                        [
+                            [
+                                -103.17130040113868,
+                                40.625392709715676
+                            ],
+                            [
+                                -103.17889641706886,
+                                40.61979008921054
+                            ]
+                        ]
+                    ],
+                },
+            ],
+            "streetNameFrom": "US-34",
+            "directionOfTraffic": " East/West ",
+            "requestedTemporarySpeed": "45",
+            "workStartDate": "2021-08-09T13:00:00.000Z",
+            "workEndDate": "2021-09-24T23:00:00.000Z",
+            "reductionJustification": "Crews roadside.",
+        }
+    }
+    assert navjoy_568.validate_closure(event) == False
+
+
+def test_validate_closure_invalid_start_time():
+    event = {
+        "sys_gUid": "Form568-cb0fdaf0-c27a-4bef-aabd-442615dfb2d6",
+        "data": {
+            "descriptionForProject": "Bridge Repairs",
+            "srzmap": [
+                {
+                    "type": "LineString",
+                    "coordinates": [
+                        [
+                            [
+                                -103.17130040113868,
+                                40.625392709715676
+                            ],
+                            [
+                                -103.17889641706886,
+                                40.61979008921054
+                            ]
+                        ]
+                    ],
+                },
+            ],
+            "streetNameFrom": "US-34",
+            "directionOfTraffic": " East/West ",
+            "requestedTemporarySpeed": "45",
+            "workStartDate": 1713004011,
+            "workEndDate": "2021-09-24T23:00:00.000Z",
+            "reductionJustification": "Crews roadside.",
+        }
+    }
+    assert navjoy_568.validate_closure(event) == False
+
+
+def test_validate_closure_invalid():
+    event = 'invalid output'
+    assert navjoy_568.validate_closure(event) == False
+
+
+def test_validate_closure_no_data():
+    event = None
+    assert navjoy_568.validate_closure(event) == False
+
+
+def test_validate_closure_no_coordinates():
+    event = {
+        "sys_gUid": "Form568-cb0fdaf0-c27a-4bef-aabd-442615dfb2d6",
+        "data": {
+            "descriptionForProject": "Bridge Repairs",
+            "srzmap": [
+                {
+                    "type": "Point",
+                    "coordinates":
+                        [
+                            -103.17130040113868,
+                            40.625392709715676
+                        ],
+                },
+            ],
+            "streetNameFrom": "US-34",
+            "directionOfTraffic": " East/West ",
+            "requestedTemporarySpeed": "45",
+            "workStartDate": "2021-08-09T13:00:00.000Z",
+            "workEndDate": "2021-09-24T23:00:00.000Z",
+            "reductionJustification": "Crews roadside.",
+        }
+    }
+    assert navjoy_568.validate_closure(event) == False
+
+
+# ----------------------------------------- get_directions_from_string -----------------------------------------
+def test_get_directions_from_string_valid():
+    directions_string = ' East/West '
+    expected = ['eastbound', 'westbound']
+    actual = navjoy_568.get_directions_from_string(directions_string)
+
+    assert actual == expected
+
+
+def test_get_directions_from_string_invalid():
+    directions_string = ' Easasdt/dWest '
+    expected = []
+    actual = navjoy_568.get_directions_from_string(directions_string)
+
+    assert actual == expected
+
+
+def test_get_directions_from_string_none():
+    directions_string = None
+    expected = []
+    actual = navjoy_568.get_directions_from_string(directions_string)
+
+    assert actual == expected
+
+
+def test_get_directions_from_string_empty_string():
+    directions_string = ''
+    expected = []
+    actual = navjoy_568.get_directions_from_string(directions_string)
+
+    assert actual == expected
+
+
+# ----------------------------------------- get_linestring_index -----------------------------------------
+def test_get_linestring_index_valid():
+    map = [
+        {
+            "type": "LineString",
+            "coordinates": [
+                [
+                    [
+                        -103.17130040113868,
+                        40.625392709715676
+                    ],
+                    [
+                        -103.17889641706886,
+                        40.61979008921054
+                    ]
+                ]
+            ],
+        },
+    ]
+    expected = 0
+    actual = navjoy_568.get_linestring_index(map)
+
+    assert actual == expected
+
+
+def test_get_linestring_no_linestring():
+    map = [
+        {
+            "type": "Polygon",
+            "coordinates": [
+                [
+                    [
+                        -103.17130040113868,
+                        40.625392709715676
+                    ],
+                    [
+                        -103.17889641706886,
+                        40.61979008921054
+                    ]
+                ]
+            ],
+        },
+    ]
+    expected = None
+    actual = navjoy_568.get_linestring_index(map)
+
+    assert actual == expected
+
+
+# ----------------------------------------- get_polygon_index -----------------------------------------
+def test_get_polygon_index_valid():
+    map = [
+        {
+            "type": "Polygon",
+            "coordinates": [
+                [
+                    [
+                        -103.17130040113868,
+                        40.625392709715676
+                    ],
+                    [
+                        -103.17889641706886,
+                        40.61979008921054
+                    ]
+                ]
+            ],
+        },
+    ]
+    expected = 0
+    actual = navjoy_568.get_polygon_index(map)
+
+    assert actual == expected
+
+
+def test_get_polygon_no_polygon():
+    map = [
+        {
+            "type": "Linestring",
+            "coordinates": [
+                [
+                    [
+                        -103.17130040113868,
+                        40.625392709715676
+                    ],
+                    [
+                        -103.17889641706886,
+                        40.61979008921054
+                    ]
+                ]
+            ],
+        },
+    ]
+    expected = None
+    actual = navjoy_568.get_polygon_index(map)
+
+    assert actual == expected
+
+
+def test_expand_speed_zone_1():
+    event = {
+        "sys_gUid": "Form568-cb0fdaf0-c27a-4bef-aabd-442615dfb2d6",
+        "data": {
+            "srzmap": [
+                {
+                    "type": "LineString",
+                    "coordinates": [
+                        [
+                            [
+                                -103.17130040113868,
+                                40.625392709715676
+                            ],
+                            [
+                                -103.17889641706886,
+                                40.61979008921054
+                            ]
+                        ]
+                    ],
+                },
+            ],
+            "streetNameFrom": "US-34",
+            "directionOfTraffic": " East/West ",
+            "requestedTemporarySpeed": "45",
+            "workStartDate": "2021-08-09T13:00:00.000Z",
+            "workEndDate": "2021-09-24T23:00:00.000Z",
+            "reductionJustification": "Crews roadside.",
+            'currentPostedSpeed': None,
+            'mileMarkerEnd': None,
+            'mileMarkerStart': None,
+        }
+    }
+
+    actual = navjoy_568.expand_speed_zone(event)
+
+    assert expected_results.test_expand_speed_zone_1_expected == actual
+
+
+def test_expand_speed_zone_2():
+    event = {
+        "sys_gUid": "Form568-cb0fdaf0-c27a-4bef-aabd-442615dfb2d6",
+        "data": {
+            "srzmap": [
+                {
+                    "type": "LineString",
+                    "coordinates": [
+                        [
+                            [
+                                -103.17130040113868,
+                                40.625392709715676
+                            ],
+                            [
+                                -103.17889641706886,
+                                40.61979008921054
+                            ]
+                        ]
+                    ],
+                },
+            ],
+            "reductionJustification": "Crews roadside.",
+
+            "streetNameFrom2": "US-34",
+            "directionOfTraffic2": " East/West ",
+            "requestedTemporarySpeed2": "45",
+            "workStartDate2": "2021-08-09T13:00:00.000Z",
+            "workEndDate2": "2021-09-24T23:00:00.000Z",
+            'currentPostedSpeed2': None,
+            'mileMarkerEnd2': None,
+            'mileMarkerStart2': None,
+        }
+    }
+
+    actual = navjoy_568.expand_speed_zone(event)
+
+    assert expected_results.test_expand_speed_zone_2_expected == actual
+
+
+# I hate how long this test is, but this is what is has to be to test all 4 at the same time
+def test_expand_speed_zone_2_3_4():
+    event = {
+        "sys_gUid": "Form568-cb0fdaf0-c27a-4bef-aabd-442615dfb2d6",
+        "data": {
+            "srzmap": [
+            ],
+            "reductionJustification": "Crews roadside.",
+            "streetNameFrom": "US-34",
+            "directionOfTraffic": " East/West ",
+            "requestedTemporarySpeed": "1",
+            "workStartDate": "2021-08-09T13:00:00.000Z",
+            "workEndDate": "2021-09-24T23:00:00.000Z",
+            'currentPostedSpeed': None,
+            'mileMarkerEnd': None,
+            'mileMarkerStart': None,
+
+            "streetNameFrom2": "US-34",
+            "directionOfTraffic2": " East/West ",
+            "requestedTemporarySpeed2": "2",
+            "workStartDate2": "2021-08-09T13:00:00.000Z",
+            "workEndDate2": "2021-09-24T23:00:00.000Z",
+            'currentPostedSpeed2': None,
+            'mileMarkerEnd2': None,
+            'mileMarkerStart2': None,
+
+            "streetNameFrom3": "US-34",
+            "directionOfTraffic3": " East/West ",
+            "requestedTemporarySpeed3": "3",
+            "workStartDate3": "2021-08-09T13:00:00.000Z",
+            "workEndDate3": "2021-09-24T23:00:00.000Z",
+            'currentPostedSpeed3': None,
+            'mileMarkerEnd3': None,
+            'mileMarkerStart3': None,
+
+            "streetNameFrom4": "US-34",
+            "directionOfTraffic4": " East/West ",
+            "requestedTemporarySpeed4": "4",
+            "workStartDate4": "2021-08-09T13:00:00.000Z",
+            "workEndDate4": "2021-09-24T23:00:00.000Z",
+            'currentPostedSpeed4': None,
+            'mileMarkerEnd4': None,
+            'mileMarkerStart4': None,
+        }
+    }
+
+    actual = navjoy_568.expand_speed_zone(event)
+
+    assert expected_results.test_expand_speed_zone_2_3_4_expected == actual
